Fix: resolve data pipeline failures - make KASSAL_API_KEY optional and fix backfill dtype
- Make kassal_api_key optional (default empty string) so indexer can run without it
- Add explicit float conversion to backfill DataFrame to prevent object dtype errors with NaN values
- Resolves 'Field required' validation error in Daily Price Scraper workflow
- Resolves 'DataFrame.dtypes' ValueError in Monthly Retrain backfill step

+++ model/backfill_predictions.py
@@ -103,7 +103,7 @@ async def run(from_month: str = "2010-01", dry_run: bool = False) -> None:
         run_date = target_month  # use 1st of month as run_date for backfill
 
         feat = {col: (None if pd.isna(row.get(col)) else float(row[col])) for col in FEATURE_COLS}
-        X = pd.DataFrame([feat])[FEATURE_COLS].fillna(0.0)
+        X = pd.DataFrame([feat])[FEATURE_COLS].fillna(0.0).astype(float)
 
         point = float(model.predict(X)[0])
         ci_lo, ci_hi = _bootstrap_ci(model, X)
